modified:   r-daemon 	modified:   r-daemon.c 	modified:   self 	modified:   self.c

+++ r-daemon.c
@@ -27,7 +27,7 @@ int write_file(char fnp[256],long int d) {
 	fclose(fp);
 }
 
-int read_file(char fnp[256]) {
+long int read_file(char fnp[256]) {
 	
 	FILE *fp;
 	char fn[256]="/tmp/";
@@ -40,6 +40,7 @@ int read_file(char fnp[256]) {
 		fscanf(fp,"%8d",&d);
 		fclose(fp);
 	} else d=0;
+	
 	return d;
 }
